+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## 2026-01-02
+
+- Added field `with_partials` to `sf.bstream.v1.BlockRequest` protobuf message: blockstream no longer sends partial blocks by default.
+
 ## 2025-12-12
 
 ### Added